Crate rustc_trans_utils

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Note

This API is completely unstable and subject to change.

Modules

collector 	 [
Experimental
]

link 	 [
Experimental
]

monomorphize 	 [
Experimental
]

trans_crate 	 [
Experimental
]
The Rust compiler.

trans_item 	 [
Experimental
]
Walks the crate looking for items/impl-items/trait-items that have either a rustc_symbol_name or rustc_item_path attribute and generates an error giving, respectively, the symbol name or item-path. This is used for unit testing the code that generates paths etc in all kinds of annoying scenarios.

Functions

check_for_rustc_errors_attr 	 [
Experimental
]
check for the #[rustc_error] annotation, which forces an error in trans. This is used to write compile-fail tests that actually test that compilation succeeds without reporting an error.

find_exported_symbols 	 [
Experimental
]
The context provided lists a set of reachable ids as calculated by middle::reachable, but this contains far more ids and symbols than we're actually exposing from the object file. This function will filter the set in the context to the set of ids which correspond to symbols that are exposed from the object file being generated.