Function rustc_trans_utils::find_exported_symbols

pub fn find_exported_symbols(tcx: TyCtxt) -> NodeSet

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The context provided lists a set of reachable ids as calculated by middle::reachable, but this contains far more ids and symbols than we're actually exposing from the object file. This function will filter the set in the context to the set of ids which correspond to symbols that are exposed from the object file being generated.

This list is later used by linkers to determine the set of symbols needed to be exposed from a dynamic library and it's also encoded into the metadata.