Trait rustc_trans_utils::trans_crate::TransCrate

pub trait TransCrate {
    type MetadataLoader: MetadataLoaderTrait;
    type OngoingCrateTranslation;
    type TranslatedCrate;
    fn metadata_loader() -> Box<MetadataLoaderTrait>;

    fn provide_local(_providers: &mut Providers);

    fn provide_extern(_providers: &mut Providers);

    fn trans_crate<'a, 'tcx>(
        tcx: TyCtxt<'a, 'tcx, 'tcx>, 
        rx: Receiver<Box<Any + Send>>
    ) -> Self::OngoingCrateTranslation;

    fn join_trans(
        trans: Self::OngoingCrateTranslation, 
        sess: &Session, 
        dep_graph: &DepGraph
    ) -> Self::TranslatedCrate;

    fn link_binary(
        sess: &Session, 
        trans: &Self::TranslatedCrate, 
        outputs: &OutputFilenames
    );

    fn dump_incremental_data(trans: &Self::TranslatedCrate);
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Associated Types

type MetadataLoader: MetadataLoaderTrait

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

type OngoingCrateTranslation

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

type TranslatedCrate

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn metadata_loader() -> Box<MetadataLoaderTrait>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn provide_local(_providers: &mut Providers)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn provide_extern(_providers: &mut Providers)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn trans_crate<'a, 'tcx>(
tcx: TyCtxt<'a, 'tcx, 'tcx>,
rx: Receiver<Box<Any + Send>>
) -> Self::OngoingCrateTranslation

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn join_trans(
trans: Self::OngoingCrateTranslation,
sess: &Session,
dep_graph: &DepGraph
) -> Self::TranslatedCrate

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn link_binary(
sess: &Session,
trans: &Self::TranslatedCrate,
outputs: &OutputFilenames
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn dump_incremental_data(trans: &Self::TranslatedCrate)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?