Module rustc_trans_utils::trans_item

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Walks the crate looking for items/impl-items/trait-items that have either a rustc_symbol_name or rustc_item_path attribute and generates an error giving, respectively, the symbol name or item-path. This is used for unit testing the code that generates paths etc in all kinds of annoying scenarios.

Structs

DefPathBasedNames 	 [
Experimental
]
Same as unique_type_name() but with the result pushed onto the given output parameter.

Enums

InstantiationMode 	 [
Experimental
]
Describes how a translation item will be instantiated in object files.

TransItem 	 [
Experimental
]

Traits

TransItemExt 	 [
Experimental
]

Functions

linkage_by_name 	 [
Experimental
]