Trait rustc_trans_utils::trans_item::TransItemExt

pub trait TransItemExt<'a, 'tcx>: Debug {
    fn as_trans_item(&self) -> &TransItem<'tcx>;

    fn instantiation_mode(
        &self, 
        tcx: TyCtxt<'a, 'tcx, 'tcx>
    ) -> InstantiationMode { ... }

    fn explicit_linkage(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> Option<Linkage> { ... }

    fn is_instantiable(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> bool { ... }

    fn to_string(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> String { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn as_trans_item(&self) -> &TransItem<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn instantiation_mode(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> InstantiationMode

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn explicit_linkage(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> Option<Linkage>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_instantiable(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns whether this instance is instantiable - whether it has no unsatisfied predicates.

In order to translate an item, all of its predicates must hold, because otherwise the item does not make sense. Type-checking ensures that the predicates of every item that is used by a valid item do hold, so we can rely on that.

However, we translate collector roots (reachable items) and functions in vtables when they are seen, even if they are not used, and so they might not be instantiable. For example, a programmer can define this public function:

pub fn foo<'a>(s: &'a mut ()) where &'a mut (): Clone {
    <&mut () as Clone>::clone(&s);
}

That function can't be translated, because the method <&mut () as Clone>::clone does not exist. Luckily for us, that function can't ever be used, because that would require for &'a mut (): Clone to hold, so we can just not emit any code, or even a linker reference for it.

Similarly, if a vtable method has such a signature, and therefore can't be used, we can just not emit it and have a placeholder (a null pointer, which will never be accessed) in its place.

fn to_string(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?